api/requirements.txt
# Python packages needed to be installed for functionality with API
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
python-dotenv==1.0.1
waitress==2.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
bcrypt==4.1.3
PyJWT==2.8.0